% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibMS.R
\name{calibMS1}
\alias{calibMS1}
\title{Convert CFM file to MGF
some stufff}
\usage{
calibMS1(filein, lmzcalib, dppm = 20, bw = 2.5, minpts = 11,
  bkpoint = median(calInt), calInt = c(80, 350))
}
\arguments{
\item{filein}{Path to the input file}

\item{lmzcalib}{Calibration masses}

\item{dppm}{Mass error}

\item{bw}{Bandwidth to group masses}

\item{minpts}{Minimum of points}

\item{bkpoint}{breakpoint mass to calculate robust regression}

\item{calInt}{calibration interval to compute}
}
\value{
xcms
}
