% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMUtils.R
\name{CFM}
\alias{CFM}
\alias{Cfm2Mgf}
\alias{ManyCfm2Mgf}
\title{Convert CFM file to MGF
some stufff}
\usage{
Cfm2Mgf(ifile, itmp, minpk = 3)

ManyCfm2Mgf(mf2exp, what = "Neg", fileout = NULL,
  root = "/media/david/ef81e301-125c-45e8-8e74-9ed5bbd3fd3b/CFM")
}
\arguments{
\item{ifile}{Path to the input file}

\item{itmp}{Path to the input file}

\item{minpk}{Path to the input file}

\item{mf2exp}{Data frame}

\item{what}{Path to the input file}

\item{fileout}{Path to the input file}

\item{root}{Root file}
}
\value{
List of list

List of list
}
