% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMUtils.R
\name{Cfm2Mgf}
\alias{Cfm2Mgf}
\title{Convert CFM file to MGF}
\usage{
Cfm2Mgf(ifile, itmp, minpk = 3)
}
\arguments{
\item{ifile}{Path to the input file}

\item{itmp}{Path to the input file}

\item{minpk}{Path to the input file}
}
\value{
List of list
}
\description{
This function loads a
}
