% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TremoloUtils.R
\name{cleanTremolo}
\alias{cleanTremolo}
\title{Clean up Tremolo mess}
\usage{
cleanTremolo(root, removeResfile = TRUE)
}
\arguments{
\item{root}{Directory to clean}

\item{removeResfile}{Should the result file be removed}
}
\value{
Nothing
}
\description{
Clean up Tremolo mess
}
