% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSMS.R
\name{combMSMS}
\alias{combMSMS}
\title{Return MS/MS sparse matrix from a list spectra}
\usage{
combMSMS(speclist, bin = 0.01, binwin = 0.1, reso = 17500,
  intlim = 10^-5, intlim2 = 200, nmax = NA, mzlim = c(-Inf, Inf))
}
\arguments{
\item{speclist}{List of spectra matrices}

\item{bin}{m/z binning}

\item{binwin}{m/z window}

\item{reso}{MS/MS resolution}

\item{intlim}{Limit on the intensity normalised to the most intense peak}

\item{intlim2}{Limit on the intensity as raw}

\item{nmax}{number of rows in the sparse matrix, i.e. max(mz)/bin}
}
\description{
Return MS/MS sparse matrix from a list spectra
}
