% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compSimMS.R
\name{compOneSimCosM}
\alias{compOneSimCosM}
\title{cosine similarity}
\usage{
compOneSimCosM(munk, mlib, mztol = 0.1, nMatch = 3, xrankScoremat = NULL,
  idunk = "unk", idlib = "lib")
}
\arguments{
\item{munk}{unknown spectra}

\item{mlib}{library spectra}

\item{mztol}{tolerance}

\item{nMatch}{min number of matches}

\item{xrankScoremat}{xrankScoremat}

\item{idunk}{id of unknown}

\item{idlib}{id of library spectra}

\item{xrankScoremat}{xrankScoremat}
}
\value{
matrix of matches
}
\description{
cosine similarity
}
