% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSMS.R
\name{compSimSPmat}
\alias{compSimSPmat}
\title{Compute smilarity between 1 or 2 MS/MS sparse matrices}
\usage{
compSimSPmat(m1, m2 = NULL, weight = c(m = NA, n = NA), plim = 10^-5,
  useSqrt = TRUE, retPerc = TRUE)
}
\arguments{
\item{m1}{Sparse matrix}

\item{m2}{Sparse}

\item{weight}{Weight for NDP c(m=0.5,n=1.5)}

\item{plim}{set anything below plim to zero}

\item{useSqrt}{should sqrt tranform the intensities}

\item{retPerc}{return perc overlap}
}
\description{
Compute smilarity between 1 or 2 MS/MS sparse matrices
}
