% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidasUtils.R
\name{dda2Midas}
\alias{dda2Midas}
\title{Interface to Midas}
\usage{
dda2Midas(obj, precmzrange = NULL, lsp2exp = NULL, limint = 10^-5,
  tempdir = "./", basenam = paste0("tmp", round(runif(1) * 1000)))
}
\arguments{
\item{obj}{dda metaboGoS object}

\item{precmzrange}{range prec m/z}

\item{lsp2exp}{names of the MS/MS spectra}

\item{limint}{Minimum relative intensity}

\item{tempdir}{Directory containing the output}

\item{basenam}{result file name}
}
\value{
List of tempdir/file locations
}
\description{
Interface to Midas
}
