% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eicXtract.R
\name{eicGather}
\alias{eicGather}
\title{Merge ROIs/EICs from several files}
\usage{
eicGather(files, eics, conv2df = FALSE)
}
\arguments{
\item{files}{list of files}

\item{eics}{ROI/EICs names to recoup across files}
}
\description{
Merge ROIs/EICs from several files
}
