% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eicXtract.R
\name{eicGather2}
\alias{eicGather2}
\title{Merge ROIs/EICs from several files}
\usage{
eicGather2(fileres, listRois, parDeco, lsamp2use,
  minHeight = parDeco$minHeightMS1, doPlot = T)
}
\arguments{
\item{listRois}{ROI/EICs names to recoup across files}

\item{parDeco}{par deco}

\item{lsamp2use}{sample to use to compute stats}

\item{minHeight}{minimum peak height in lsamp2use}

\item{doPlot}{should it be plotted}

\item{files}{list of files}
}
\description{
Merge ROIs/EICs from several files
}
