% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eicXtract.R
\name{eicXtract}
\alias{eicXtract}
\title{refine PI ROI

Refine/integra ROI around spectra  after association of spectra to ROIs from potential precursors}
\usage{
eicXtract(infile, eicmat, addppm = NA, refFct = metaboGoS:::.MGfill, drt,
  nSlaves = 1, save = FALSE, outfile = NULL, paddRT = drt, ...)
}
\arguments{
\item{infile}{metaboGoS obj}

\item{eicmat}{eicmat}

\item{addppm}{add ppm}

\item{refFct}{refining function}

\item{drt}{plotting at each step}

\item{nSlaves}{number of slaves for parallel}

\item{save}{numbl}

\item{outfile}{n}

\item{paddRT}{extra padding for RT}
}
\value{
metaboGoS obj with RoiInfos / PeakInfos /params
}
