% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{export2MSP}
\alias{export2MSP}
\title{Export a list of files to MSP}
\usage{
export2MSP(spec, outfile = NULL, plim = 10^-5, roundint = 0,
  roundmz = 5, topSp = Inf)
}
\arguments{
\item{spec}{list of spectra}

\item{outfile}{file to be written}

\item{plim}{min intensity in perc.}

\item{roundint}{rounding for intensity}

\item{roundmz}{rounding for m/z}
}
\value{
List of tempdir/file locations
}
\description{
Export a list of files to MSP
}
