% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS1.r
\name{extractPI}
\alias{extractPI}
\title{Extract ROI around spectra

some stufff}
\usage{
extractPI(obj, parDeco = list(minPurity = c(0.2, 0.4), ppm = 5, dmz = 0.001,
  drt = 1, rtlim = c(-Inf, -Inf, Inf, Inf), minHeightMS1 = 1e+05, minNoiseMS1 =
  10000, minZero = 200), minPIwindow = NA, PIweight = 2, nSlaves = 1)
}
\arguments{
\item{obj}{metaboGoS obj}

\item{parDeco}{deconvolution parameter list}

\item{minPIwindow}{window size in min. around the spectra}

\item{PIweight}{weighting}

\item{nSlaves}{number of slaves for parallel}
}
\value{
List of stuf
}
