% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS1.R
\name{extractPIRoi}
\alias{extractPIRoi}
\title{Extract ROI around spectra

Fast association of spectra to ROIs from potential precursors}
\usage{
extractPIRoi(obj, parDeco = list(minPurity = c(0.2, 0.4), ppm = 5, dmz =
  0.001, psdrt = NA, drt = 1, rtlim = c(-Inf, -Inf, Inf, Inf), minHeightMS1 =
  1e+05, minZero = 200), minPIwindow = NA, PIweight = 2)
}
\arguments{
\item{obj}{metaboGoS obj}

\item{parDeco}{deconvolution parameter list}

\item{minPIwindow}{window size in min. around the spectra}

\item{PIweight}{weighting}

\item{nSlaves}{number of slaves for parallel}
}
\value{
metaboGoS obj with RoiInfos / MS2ROI /params
}
