% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRoiMS1.R
\name{extractRoiMS1}
\alias{extractRoiMS1}
\title{extractRoiMS1

Refine/integra ROI}
\usage{
extractRoiMS1(samp, parDeco, blList = NULL, lmzExcl = c(), stepStop = 3,
  doPlot = TRUE, mzrange = c(-Inf, Inf), useMinNoise = TRUE)
}
\arguments{
\item{samp}{xml file or xcms object}

\item{parDeco}{deconvolution parameter list}

\item{blList}{list of blank filles xcmsRaw object}

\item{lmzExcl}{list of m/z excluded form the blank correction}

\item{doPlot}{plotting at each step}

\item{mzrange}{mzrange}

\item{useMinNoise}{useMinNoise}
}
\value{
RoiInfos / PeakInfos data.frame
}
