% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidasUtils.R
\name{genMidasCfg}
\alias{genMidasCfg}
\title{Configuration file for Midas}
\usage{
genMidasCfg(metLib, polarity = "pos", precTol = 0.1, fragTol = 0.01,
  breakRing = TRUE, fragDepth = 3, nSlaves = 1)
}
\arguments{
\item{metLib}{metabolite library}

\item{polarity}{polarity}

\item{precTol}{precursor tolerance}

\item{fragTol}{fragment tolerance}

\item{breakRing}{break Ring TRUE/FALSE}

\item{fragDepth}{result file name}

\item{nSlaves}{num of CPU}
}
\value{
string to print out
}
\description{
Configuration file for Midas
}
