% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TremoloUtils.R
\name{getTremololist}
\alias{getTremololist}
\title{set-up Tremolo input}
\usage{
getTremololist(what, lsp2exp, file2exp = NULL, perclim = 0.1, nmin = 5,
  exec = "/media/D01/Metabo/MetSoft/Tremolo_8_21_2013/convert")
}
\arguments{
\item{what}{Object containg spectra}

\item{lsp2exp}{List of spectra}

\item{file2exp}{where to write the MGF formatted spectra}

\item{perclim}{Minimum intensity in percentage}

\item{nmin}{Minimum number of peaks higher than perclim}

\item{exec}{Conversion executable}
}
\value{
List of stuff
}
\description{
Parse Tremolo results
}
