% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.chkclosemz}
\alias{infctMRoi.chkclosemz}
\title{Merge slices from the same sample within a single ROI}
\usage{
infctMRoi.chkclosemz(tmproi, indicatorVec, thrMZ = 0.001, dmz = 5e-04,
  dppm = 2.5)
}
\arguments{
\item{tmproi}{Data frame of ROIs}

\item{thrMZ}{overlap in mz}

\item{dmz}{dmz}

\item{dppm}{dppm}

\item{sid}{Sample id indicator}

\item{mzmed}{mzmed}

\item{mzmin}{mzmin}

\item{mzmax}{mzmax}
}
\value{
list of entries to be merged in tmproi - may be null
}
\description{
Merge slices from the same sample within a single ROI
}
\keyword{internal}
