% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.get}
\alias{infctMRoi.get}
\title{Gather the rois for each resfile}
\usage{
infctMRoi.get(fileres, what = "RoiInfos")
}
\arguments{
\item{fileres}{List of files}

\item{what}{what}
}
\description{
Gather the rois for each resfile
}
\keyword{internal}
