% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.loopmzrt}
\alias{infctMRoi.loopmzrt}
\title{Iterative merging of slices with overlaping mz/rt within a single ROI regardless sample Id}
\usage{
infctMRoi.loopmzrt(tmproi, indicatorVec, thrMZ = 0.001, thrRT = 1e-04,
  maxiter = 6)
}
\arguments{
\item{tmproi}{Data frame of ROIs}

\item{indicatorVec}{column indicator vector}

\item{thrMZ}{overlap in mz}

\item{thrRT}{overlap in rt}

\item{maxiter}{maximum iterations, usually done within 1-2}
}
\value{
list of entries in tmproi
}
\description{
Iterative merging of slices with overlaping mz/rt within a single ROI regardless sample Id
}
\keyword{internal}
