% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.mergeclosemz}
\alias{infctMRoi.mergeclosemz}
\title{Wrapper for infctMRoi.chkclosemz: loop over GrpEIC}
\usage{
infctMRoi.mergeclosemz(currrois, currstats, indicatorVec, thrMZ = 0.001,
  dmz = 5e-04, dppm = 2.5)
}
\arguments{
\item{currrois}{Data frame of ROIs}

\item{currstats}{Stats}

\item{thrMZ}{overlap in mz}

\item{dmz}{dmz}

\item{dppm}{dppm}
}
\description{
Wrapper for infctMRoi.chkclosemz: loop over GrpEIC
}
\keyword{internal}
