% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.mergemzrt}
\alias{infctMRoi.mergemzrt}
\title{1 Merge based on dmz or overlap in mz 2 merge on mz/rt  wrapper for infctMRoi.loopmzrt}
\usage{
infctMRoi.mergemzrt(currrois, currstats, typ = "mzmed", indicatorVec,
  dppm = 2.5, dmz = 5e-04, thrMZ = 0.001, thrRT = 0, maxiter = 6)
}
\arguments{
\item{currrois}{Data frame of ROIs}

\item{currstats}{Sample id indicator}

\item{typ}{mz column indicator}

\item{dppm}{dppm}

\item{dmz}{dmz}

\item{thrMZ}{overlap in mz}

\item{mzmin}{mzmin}

\item{mzmax}{mzmax}
}
\value{
list of entries to be merged in tmproi - may be null
}
\description{
1 Merge based on dmz or overlap in mz 2 merge on mz/rt  wrapper for infctMRoi.loopmzrt
}
\keyword{internal}
