% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.splitmzdens}
\alias{infctMRoi.splitmzdens}
\title{Split ROI based on density mz}
\usage{
infctMRoi.splitmzdens(crois, cstats, lurois, indicatorVec, bw = 5,
  cid = max(cstats[, indicatorVec["grp"]]))
}
\arguments{
\item{crois}{ROIs}

\item{cstats}{Stats}

\item{lurois}{vector of grp ids be splitted}

\item{indicatorVec}{indicatorVec}

\item{bw}{ppm bandwidth parameter}

\item{cid}{Group EIC numeric id}
}
\value{
list ROI/stats after splitting crois -> replace crois/cstats
}
\description{
Split ROI based on density mz
}
\keyword{internal}
