% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeROI.r
\name{infctMRoi.splitmzrt}
\alias{infctMRoi.splitmzrt}
\title{Double merge on mz/rt - wrapper for infctMRoi.loopmzrt}
\usage{
infctMRoi.splitmzrt(crois, cstats, indicatorVec, lurois = unique(cstats[,
  indicatorVec["grp"]]), cid = max(cstats[, indicatorVec["grp"]]),
  thrMZ = 0.001, thrRT = 0.001)
}
\arguments{
\item{indicatorVec}{indicator}

\item{lurois}{vector of rois}

\item{cid}{grp max value to start from}

\item{thrMZ}{overlap in mz}

\item{thrRT}{overlap in rt}

\item{currrois}{Data frame of ROIs}

\item{currstats}{Sample id indicator}
}
\value{
list of entries to be merged in tmproi - may be null
}
\description{
Double merge on mz/rt - wrapper for infctMRoi.loopmzrt
}
\keyword{internal}
