% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMUtils.R
\name{ManyCfm2Mgf}
\alias{ManyCfm2Mgf}
\title{Multiple CFM file to MGF}
\usage{
ManyCfm2Mgf(mf2exp, what = "Neg", fileout = NULL,
  root = "/media/david/ef81e301-125c-45e8-8e74-9ed5bbd3fd3b/CFM")
}
\arguments{
\item{mf2exp}{Data frame}

\item{what}{Path to the input file}

\item{fileout}{Path to the input file}

\item{root}{Root file}
}
\value{
List of list
}
\description{
This function loads a
}
