% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboSet.R
\name{mapMS2toMetaboSet}
\alias{mapMS2toMetaboSet}
\title{mapMS2toMetaboSet

stuff}
\usage{
mapMS2toMetaboSet(cmb, fileres, dmz = 0.001, dppm = 2, rtwin = 3/60,
  outfile = NULL)
}
\arguments{
\item{cmb}{metaboSet object}

\item{fileres}{list DDA files}

\item{dmz}{delta mz}

\item{dppm}{delta ppm}

\item{rtwin}{rtwin}

\item{outfile}{file to be exported}
}
\value{
metaboSet
}
