% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSMS.R
\name{mergeMSMS.2}
\alias{mergeMSMS.2}
\title{Merge list of ms/ms spectra}
\usage{
mergeMSMS.2(splist, dmz = 0.001, fct = sum)
}
\arguments{
\item{dmz}{delta m/z for merging}

\item{fct}{merging function}

\item{split}{list of spectra}
}
\description{
Merge list of ms/ms spectra
}
