% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSMS.R
\name{mergeMSMS}
\alias{mergeMSMS}
\title{Merge list of ms/ms spectra}
\usage{
mergeMSMS(splist, dmz = 0.005, fct = median, minNoise = 0, reso = 17500)
}
\arguments{
\item{dmz}{delta m/z for merging}

\item{fct}{merging function}

\item{minNoise}{supress anything below minNoise}

\item{split}{list of spectra}
}
\description{
Merge list of ms/ms spectra
}
