% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDDA.R
\name{parseOneDDA}
\alias{parseOneDDA}
\title{Load DDA run from a mzXML file}
\usage{
parseOneDDA(cefi, parDeco = NULL, add2File = NULL, defIsolation = 0.501,
  calibFct = NULL, cleanMS2 = TRUE, save = FALSE, outfile = NULL,
  roundmz = 6, verbose = T, sid = add2File$Sid, useSid = !is.null(sid),
  stype = NULL, rtlim = suppressWarnings(range(parDeco$rtlim)))
}
\arguments{
\item{cefi}{Path to the input file}

\item{parDeco}{List of decoParams}

\item{add2File}{Additional information to complement the file infos}

\item{defIsolation}{Default isolation window around precursor}

\item{save}{Save FALSE/TRUE}

\item{outfile}{Filename to save}

\item{roundmz}{Rounding for the mass}

\item{verbose}{Print stuff out}

\item{sid}{Sample id to given}

\item{useSid}{Add sid to the spectra names}

\item{stype}{Sample type}

\item{rtlim}{Limit MS/MS to a certain range}
}
\value{
List of list
}
\description{
This function loads a DDA/DIA/AIF run from a mzXML file
}
