% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDDA.R
\name{plotOneDDA}
\alias{plotOneDDA}
\title{Plot DDA object}
\usage{
plotOneDDA(sfile, blfile = NULL, dmz = 0.001, dppm = 5, drt = 1,
  winMS2 = 0.501)
}
\arguments{
\item{sfile}{sfile}

\item{blfile}{blfile}

\item{dmz}{dmz}

\item{dppm}{dppm}

\item{drt}{drt}

\item{winMS2}{winMS2}
}
\description{
Plot DDA object
}
\keyword{internal}
