% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compSimMS.R
\name{plotResMatch}
\alias{plotResMatch}
\title{plotting}
\usage{
plotResMatch(resMatch, type = 1)
}
\arguments{
\item{resMatch}{result from compOneSimCosM match}

\item{typ}{1/2}
}
\value{
matrix of matches
}
\description{
plotting
}
