% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{readMGFfile}
\alias{readMGFfile}
\title{Read MGF file}
\usage{
readMGFfile(filein, namIdx = "DBSpId", startIdx = 1, minpk = 1)
}
\arguments{
\item{filein}{list of files}

\item{namIdx}{append to spec nam}

\item{startIdx}{starting number for spectrum name}
}
\value{
List of tempdir/file locations
}
\description{
Read MGF file
}
