% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS1.R
\name{refinePIRoi}
\alias{refinePIRoi}
\title{refine PI ROI

Refine/integra ROI around spectra  after association of spectra to ROIs from potential precursors}
\usage{
refinePIRoi(obj, parDeco, refFct = metaboGoS:::.MGrefFct0, doPlot = TRUE,
  blList = NULL, nSlaves = 1, useMS1file = TRUE)
}
\arguments{
\item{obj}{metaboGoS obj}

\item{parDeco}{deconvolution parameter list}

\item{refFct}{refining function}

\item{doPlot}{plotting at each step}

\item{blList}{list of blank filles xcmsRaw object}

\item{nSlaves}{number of slaves for parallel}

\item{useMS1file}{use the MS1 scan in MS1file}
}
\value{
metaboGoS obj with RoiInfos / PeakInfos /params
}
