% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TremoloUtils.R
\name{runTremolo}
\alias{runTremolo}
\title{read Tremolo output}
\usage{
runTremolo(root, speclib, prectol = 0.1, score = 0.4, topk = 10,
  exec = "/media/D01/Metabo/MetSoft/Tremolo_8_21_2013/main_execmodule")
}
\arguments{
\item{root}{Directory to store the results}

\item{speclib}{Spectra lib in .mgf format}

\item{prectol}{Precursor tolerance}

\item{score}{Minimum score}

\item{topk}{Only keep the top matches}

\item{exec}{Tremolo execuytable}
}
\value{
Data frame
}
\description{
Parse Tremolo results
}
