% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsUtil.R
\name{satremoval}
\alias{satremoval}
\title{Artfact removal in duples}
\usage{
satremoval(re, reso, typ = 2)
}
\arguments{
\item{re}{matrix of mz/int}

\item{reso}{mass resolution}

\item{typ}{1/2 extended}
}
\value{
ids of entries to be remove
}
\description{
Artfact removal in duples
}
