% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsUtil.R
\docType{data}
\name{slimXcmsRaw}
\alias{slimXcmsRaw}
\alias{slimXcmsRaw,xcmsRaw-method}
\alias{slimXcmsRaw}
\title{get the xcmsRaw lighter}
\format{An object of class \code{NULL} of length 0.}
\usage{
\S4method{slimXcmsRaw}{xcmsRaw}(object, eicmat, inthr = 0)
}
\description{
get the xcmsRaw lighter
}
\keyword{datasets}
