% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibMS.R
\name{calibMS1}
\alias{calibMS1}
\title{Convert CFM file to MGF
some stufff}
\usage{
calibMS1(filein, lmzcalib, dppm = 20, bw = 2.5, minpts = 31,
  bkpoint = median(calInt), calInt = c(70, 500), typ = 1, doPlot = T,
  retMat = F)
}
\arguments{
\item{filein}{Path to the input file or xcmsRaw object}

\item{lmzcalib}{Calibration masses}

\item{dppm}{Mass error}

\item{bw}{Bandwidth to group masses}

\item{minpts}{Minimum of points}

\item{bkpoint}{breakpoint mass to calculate robust regression}

\item{calInt}{calibration interval to compute}

\item{doPlot}{should plot}
}
\value{
xcms
}
