% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFormUtils.R
\name{getMF.7GR}
\alias{getMF.7GR}
\title{Interface to the 7GR executable}
\usage{
getMF.7GR(precmz, lIP = "[M+H]1+", precppm = 11, precmtol = 10,
  ipdb = NULL, exec = "/media/D01/Metabo/MetSoft/HR/myhr", chkdb = NULL)
}
\arguments{
\item{precmz}{Precursor m/z}

\item{lIP}{Vector of ionisation products}

\item{precppm}{Accuracy in ppm}

\item{precmtol}{Tolerence in mDa}

\item{ipdb}{IP database}

\item{exec}{7GR executable}

\item{chkdb}{database of molform}
}
\value{
List of stuff
}
\description{
Interface to 7GR executable
}
