% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDDA.R
\name{parseOneDDA}
\alias{parseOneDDA}
\title{Load DDA run from a mzXML file}
\usage{
parseOneDDA(cefi, sid = NULL, add2File = NULL, save = FALSE,
  outfile = NULL, stype = NA, roundmz = 6, verbose = T,
  useSid = FALSE, rtlim = c(-Inf, Inf), MinIntBP = 1000)
}
\arguments{
\item{cefi}{Path to the input file}

\item{sid}{Sample id to given}

\item{add2File}{Additional information to complement the file infos}

\item{save}{Save FALSE/TRUE}

\item{outfile}{Filename to save}

\item{stype}{Sample type}

\item{roundmz}{Rounding for the mass}

\item{verbose}{Print stuff out}

\item{useSid}{Add sid to the spectra names}

\item{rtlim}{Limit MS/MS to a certain range}

\item{MinIntBP}{Minimum MS/MS basepeak  height}
}
\value{
List of list
}
\description{
This function loads a DDA/DIA/AIF run from a mzXML file
}
