% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsUtil.R
\docType{data}
\name{satremXcmsRaw}
\alias{satremXcmsRaw}
\alias{satremXcmsRaw,xcmsRaw-method}
\alias{satremXcmsRaw}
\title{get the xcmsRaw lighter}
\format{An object of class \code{NULL} of length 0.}
\usage{
\S4method{satremXcmsRaw}{xcmsRaw}(object, reso = 70000, typ = 2,
  inthr = 0, rtlim = c(-Inf, Inf))
}
\description{
get the xcmsRaw lighter
}
\keyword{datasets}
